refactor: rename files component to files-mongodb
