+++ app/admin/files/mongodb/page.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 
-import FilesAdmin from "@/components/files/admin";
+import FilesAdmin_MongoDB from "@/components/files-mongodb/admin";
 
 const Files: React.FC = () => {
-	return <FilesAdmin className="margin_vh_top margin_vh_bottom scroll-m-40" />;
+	return <FilesAdmin_MongoDB className="margin_vh_top margin_vh_bottom scroll-m-40" />;
 };
 
 export default Files;
