+++ components/_common.actions.ts
@@ -3,7 +3,10 @@ import { ObjectId } from "mongodb";
 import { getServerSession } from "next-auth";
 import { revalidatePath } from "next/cache";
 
-import { fileAttachment_add, fileAttachment_remove } from "@/components/files/_files.actions";
+import {
+	fileAttachment_add,
+	fileAttachment_remove,
+} from "@/components/files-mongodb/_files.actions"; // TODO: files-cloudflare
 import { tagAttachment_add, tagAttachment_remove } from "@/components/tags/_tags.actions";
 import { AboutEntryDoc, NewAboutEntryData } from "@/interfaces/AboutEntry";
 import { NewProjectData, ProjectDoc } from "@/interfaces/Project";
