+++ components/files-mongodb/admin/Actions/CreateFile.tsx
@@ -3,7 +3,7 @@ import React, { useState } from "react";
 
 import { usePathname } from "next/navigation";
 
-import { createFile } from "@/components/files/_files.actions";
+import { createFile } from "@/components/files-mongodb/_files.actions";
 import ButtonIcon from "@/components/fragments/ButtonIcon";
 import {
 	Dialog,
