+++ components/pages/admin/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { getFileList } from "@/components/files/_files.actions";
+import { getFileList } from "@/components/files-mongodb/_files.actions"; // TODO: files-cloudflare
 
 import { getPageCards } from "../_pages.actions";
 import styles from "../_pages.module.scss";
