+++ components/portfolio/admin/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { getFileList } from "@/components/files/_files.actions";
+import { getFileList } from "@/components/files-mongodb/_files.actions"; // TODO: files-cloudflare
 import { getTags } from "@/components/tags/_tags.actions";
 
 import { getProjects } from "../_portfolio.actions";
